Tests of directed separation

Evaluation of conditional independence claims to be used in determining the goodness-of-fit for piecewise structural equation models.

dSep(modelList, direction = NULL, conserve = FALSE, conditioning = FALSE,
  .progressBar = TRUE)

Arguments

modelList
A list of structural equations created using psem.

direction
A vector of claims defining the specific directionality of independence claims; for use in special cases (see Details).

conserve
Whether the most conservative P-value should be returned; for use in special cases (see Details). Default is FALSE.

conditioning
Whether the conditioning variables should be shown in the summary table. Default is FALSE.

.progressBar
An optional progress bar. Default is TRUE.

Value

Returns a data.frame of independence claims and their significance values.

Details

In cases involving non-normally distributed responses in the independence claims that are modeled using generalized linear models, the significance of the independence claim is not reversable (e.g., the P-value of Y ~ X is not the same as X ~ Y). This is due to the transformation of the response via the link function. In extreme cases, this can bias the goodness-of-fit tests. summary.psem will issue a warning when this case is present and provide guidance for solutions.

One solution is to specify the directionality of the relationship using the direction argument, e.g. direction = c("X <- Y"). Another is to run both tests (Y ~ X, X ~ Y) and return the most conservative (i.e., lowest) P-value, which can be toggled using the conserve = TRUE argument.

References

Shipley, Bill. "A new inferential test for path models based on directed acyclic graphs." Structural Equation Modeling 7.2 (2000): 206-218.

See also