Summarizing piecewise structural equation models

Returns information necessary to interpret piecewise structural equation models, including tests of directed separation, path coefficients, information criterion values, and R-squared values of individual models.

# S3 method for psem
summary(object, ..., direction = NULL, conserve = FALSE,
  conditional = FALSE, add.claims = NULL, standardize = "scale",
  standardize.type = "latent.linear", intercepts = FALSE,
  .progressBar = TRUE)

Arguments

object
a list of structural equations

...
additional arguments to summary

direction
a vector of claims defining the specific directionality of any independence claim(s)

conserve
whether the most conservative P-value should be returned (See Details) Default is FALSE

conditional
whether all conditioning variables should be shown in the table Default is FALSE

add.claims
an optional vector of additional independence claims (P-values) to be added to the basis set

standardize
whether standardized path coefficients should be reported Default is "scale"

standardize.type
the type of standardized for non-Gaussian responses: latent.linear (default), Mendard.OE

intercepts
whether intercepts should be included in the coefficient table Default is FALSE

.progressBar
an optional progress bar. Default is TRUE

Value

The function summary.psem returns a list of summary statistics:

dTable

A summary table of the tests of directed separation, from dSep.

CStat

Fisher's C statistic, degrees of freedom, and significance value based on a Chi-square test.

IC

Information criterion (Akaike, Bayesianm, corrected Akaike) as well as degrees of freedom and sample size.

coefficients

A summary table of the path coefficients, from link{coefs}.

R2

(Pseudo)-R2 values, from rsquared.

Details

The forthcomiing argument groups splits the analysis based on an optional grouping factor, conducts separate d-sep tests, and reports goodness-of-fit and path coefficients for each submodel. The procedure is approximately similar to a multigroup analysis in traditional variance-covariance SEM. Coming in version 2.1.

In cases involving non-normally distributed responses in the independence claims that are modeled using generalized linear models, the significance of the independence claim is not reversable (e.g., the P-value of Y ~ X is not the same as X ~ Y). This is due to the transformation of the response via the link function. In extreme cases, this can bias the goodness-of-fit tests. summary.psem will issue a warning when this case is present and provide guidance for solutions. One solution is to specify the directionality of the relationship using the direction argument, e.g. direction = c("X <- Y"). Another is to run both tests (Y ~ X, X ~ Y) and return the most conservative (i.e., lowest) P-value, which can be toggled using the conserve = TRUE argument.

In some cases, additional claims that were excluded from the basis set can be added back in using the argument add.claims. These could be, for instance, independence claims among exogenous variables. See Details in basisSet.

Standardized path coefficients are scaled by standard deviations.

References

Shipley, Bill. "A new inferential test for path models based on directed acyclic graphs." Structural Equation Modeling 7.2 (2000): 206-218.

Shipley, Bill. Cause and correlation in biology: a user's guide to path analysis, structural equations and causal inference. Cambridge University Press, 2002.

Shipley, Bill. "Confirmatory path analysis in a generalized multilevel context." Ecology 90.2 (2009): 363-368.

Shipley, Bill. "The AIC model selection method applied to path analytic models compared using a d-separation test." Ecology 94.3 (2013): 560-564.

See also